requirements/prod.txt
boto3==1.29.1
supabase==2.0.3
openai==1.3.0
google-cloud-translate==3.12.1
devatrans==5.0.1